ryan requests

+++ src/pynwb/ndx_fl_novela/probe.py
@@ -10,9 +10,9 @@ class ShanksElectrode(NWBContainer):
 
     @docval(
         {'name': 'name', 'type': str, 'doc': 'name of the shank'},
-        {'name': 'rel_x', 'type': int, 'doc': 'the rel_x value of this electrode'},
-        {'name': 'rel_y', 'type': int, 'doc': 'the rel_y value of this electrode'},
-        {'name': 'rel_z', 'type': int, 'doc': 'the rel_z value of this electrode'},
+        {'name': 'rel_x', 'type': float, 'doc': 'the rel_x value of this electrode'},
+        {'name': 'rel_y', 'type': float, 'doc': 'the rel_y value of this electrode'},
+        {'name': 'rel_z', 'type': float, 'doc': 'the rel_z value of this electrode'},
     )
     def __init__(self, **kwargs):
         super(ShanksElectrode, self).__init__(name=kwargs['name'])
