+++ src/spec/create_extension_spec.py
@@ -38,17 +38,17 @@ def main():
             NWBAttributeSpec(
                 name='rel_x',
                 doc='the rel_x value of this electrode',
-                dtype='int'
+                dtype='float'
             ),
             NWBAttributeSpec(
                 name='rel_y',
                 doc='the rel_y value of this electrode',
-                dtype='int'
+                dtype='float'
             ),
             NWBAttributeSpec(
                 name='rel_z',
                 doc='the rel_z value of this electrode',
-                dtype='int'
+                dtype='float'
             ),
         ]
     )
